+++ WowPacketParserModule.V6_0_2_19033/Parsers/QuestHandler.cs
@@ -544,11 +544,11 @@ namespace WowPacketParserModule.V6_0_2_19033.Parsers
             int int252 = packet.ReadInt32("EmotesCount");
 
             // QuestDescEmote
-            questOfferReward.Emote = new uint?[] {0, 0, 0, 0};
+            questOfferReward.Emote = new int?[] {0, 0, 0, 0};
             questOfferReward.EmoteDelay = new uint?[] {0, 0, 0, 0};
             for (int i = 0; i < int252; i++)
             {
-                questOfferReward.Emote[i] = (uint)packet.ReadInt32("Type");
+                questOfferReward.Emote[i] = packet.ReadInt32("Type");
                 questOfferReward.EmoteDelay[i] = packet.ReadUInt32("Delay");
             }
